requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
django-bulk-update-or-create==0.3.0
django-filter==22.1
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
